feat(react-grab): add scoped container + display-only demo modes
Adds two gated runtime modes so React Grab can run confined to a single
element without taking over the host page:

- `container`: hit-testing is filtered to the container's subtree and the
  toolbar treats the container's box as its viewport.
- `demo`: display-only showcase — real user input is ignored (synthetic
  events still drive it), the clipboard is never written, and the overlay
  is click-through.

Both share a single `runtime-mode` singleton, and every host-page side
effect (page freeze, global animation freeze, React-update freeze, cursor
override, body style writes) is now gated so scoped mode never mutates the
surrounding page. Toolbar re-anchors to the container on scroll. Adds a
`reset()` API and a canonical `DEFAULT_TOOLBAR_STATE`.

Uses these to build the website-v2 homepage: an autoplaying React Grab
demo card driven via the API, a /shadcn component preview, a light/dark
theme toggle, and an install command — rethemed via shadcn design tokens.
