+++ packages/react-grab/src/components/toolbar/state.ts
@@ -6,6 +6,14 @@ export type SnapEdge = "top" | "bottom" | "left" | "right";
 
 const STORAGE_KEY = "react-grab-toolbar-state";
 
+export const DEFAULT_TOOLBAR_STATE: ToolbarState = {
+  edge: "bottom",
+  ratio: TOOLBAR_DEFAULT_POSITION_RATIO,
+  collapsed: false,
+  enabled: true,
+  defaultAction: DEFAULT_ACTION_ID,
+};
+
 export const loadToolbarState = (): ToolbarState | null => {
   try {
     const serializedToolbarState = localStorage.getItem(STORAGE_KEY);
@@ -22,12 +30,14 @@ export const loadToolbarState = (): ToolbarState | null => {
         record.edge === "left" ||
         record.edge === "right"
           ? record.edge
-          : "bottom",
-      ratio: typeof record.ratio === "number" ? record.ratio : TOOLBAR_DEFAULT_POSITION_RATIO,
+          : DEFAULT_TOOLBAR_STATE.edge,
+      ratio: typeof record.ratio === "number" ? record.ratio : DEFAULT_TOOLBAR_STATE.ratio,
       collapsed,
       enabled: !collapsed,
       defaultAction:
-        typeof record.defaultAction === "string" ? record.defaultAction : DEFAULT_ACTION_ID,
+        typeof record.defaultAction === "string"
+          ? record.defaultAction
+          : DEFAULT_TOOLBAR_STATE.defaultAction,
     };
   } catch (error) {
     console.warn("[react-grab] Failed to load toolbar state from localStorage:", error);
