+++ packages/react-grab/src/core/index.tsx
@@ -20,6 +20,7 @@ import {
   hasTextSelectionOnPage,
 } from "../utils/is-keyboard-event-triggered-by-input.js";
 import { mountRoot } from "../utils/mount-root.js";
+import { setRuntimeMode, getScopeContainer } from "../utils/runtime-mode.js";
 import { createComponentNameForElement } from "../utils/create-component-name-for-element.js";
 import { watchAppTheme } from "../utils/detect-app-theme.js";
 import {
@@ -121,7 +122,11 @@ import { getScriptOptions } from "../utils/get-script-options.js";
 import { isEnterCode } from "../utils/is-enter-code.js";
 import { isMac } from "../utils/is-mac.js";
 import { isPositionInsideBounds } from "../utils/is-position-inside-bounds.js";
-import { loadToolbarState, saveToolbarState } from "../components/toolbar/state.js";
+import {
+  loadToolbarState,
+  saveToolbarState,
+  DEFAULT_TOOLBAR_STATE,
+} from "../components/toolbar/state.js";
 import { copyPlugin } from "./plugins/copy.js";
 import { commentPlugin } from "./plugins/comment.js";
 import { editPlugin } from "./plugins/edit.js";
@@ -203,6 +208,9 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
   }
   hasInited = true;
 
+  const isDemoMode = Boolean(initialOptions.demo);
+  setRuntimeMode(initialOptions.container ?? null, isDemoMode);
+
   logIntro(initialOptions.telemetry !== false);
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars -- omit init-only options that aren't part of SettableOptions
@@ -266,16 +274,24 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       );
     });
 
+    // Single boundary for host-page mutations: scoped mode (overlay confined to
+    // a container) must never touch the surrounding page, so every body-style
+    // write goes through here. The freeze helpers self-gate the same way.
+    const setHostBodyStyle = (property: "userSelect" | "touchAction", value: string) => {
+      if (getScopeContainer()) return;
+      document.body.style[property] = value;
+    };
+
     createEffect(
       on(isActivated, (activated, previousActivated) => {
         if (activated && !previousActivated) {
           freezePseudoStates(pointer().x, pointer().y);
           freezeGlobalAnimations();
-          document.body.style.touchAction = "none";
+          setHostBodyStyle("touchAction", "none");
         } else if (!activated && previousActivated) {
           unfreezePseudoStates();
           unfreezeGlobalAnimations();
-          document.body.style.touchAction = "";
+          setHostBodyStyle("touchAction", "");
         }
       }),
     );
@@ -854,6 +870,8 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
     createEffect(
       on(isActivated, (activated) => {
         if (!activated) return;
+        // Scoped mode must not pause the host app's React updates.
+        if (getScopeContainer()) return;
         if (!pluginRegistry.store.options.freezeReactUpdates) return;
         const unfreezeUpdates = freezeUpdates();
         onCleanup(unfreezeUpdates);
@@ -1298,6 +1316,9 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
     let cursorStyleElement: HTMLStyleElement | null = null;
 
     const setCursorOverride = (cursor: string | null) => {
+      // Scoped mode must not restyle the host page's cursor; the embedder (e.g.
+      // the demo) controls the cursor within its own container.
+      if (getScopeContainer()) return;
       if (cursor) {
         if (!cursorStyleElement) {
           cursorStyleElement = document.createElement("style");
@@ -1350,7 +1371,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       setIsPendingContextMenuSelect(false);
       setPendingToolbarActionId(null);
       if (wasDragging) {
-        document.body.style.userSelect = "";
+        setHostBodyStyle("userSelect", "");
       }
       if (keydownSpamTimerId) window.clearTimeout(keydownSpamTimerId);
       autoScroller.stop();
@@ -1674,7 +1695,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
 
       actions.startDrag({ x: clientX, y: clientY }, isShiftHeld);
       actions.setPointer({ x: clientX, y: clientY });
-      document.body.style.userSelect = "none";
+      setHostBodyStyle("userSelect", "none");
 
       scheduleDragPreviewUpdate(clientX, clientY);
 
@@ -1954,7 +1975,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       stopSpaceDragRepositioning();
       actions.cancelDrag();
       autoScroller.stop();
-      document.body.style.userSelect = "";
+      setHostBodyStyle("userSelect", "");
     };
 
     const handlePointerUp = (
@@ -1986,7 +2007,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       }
       stopSpaceDragRepositioning();
       autoScroller.stop();
-      document.body.style.userSelect = "";
+      setHostBodyStyle("userSelect", "");
 
       if (dragSelectionRect) {
         handleDragSelection(dragSelectionRect, hasModifierKeyHeld, isShiftHeld);
@@ -2342,9 +2363,19 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       }
     };
 
+    // Demo mode ignores real user input but still processes synthetic events
+    // (which drive the showcase), so the guard lives here once instead of being
+    // repeated inside every interaction handler.
+    const ignoreRealInput =
+      <EventType extends Event>(handler: (event: EventType) => void) =>
+      (event: EventType): void => {
+        if (isDemoMode && event.isTrusted) return;
+        handler(event);
+      };
+
     eventListenerManager.addWindowListener(
       "keydown",
-      (event: KeyboardEvent) => {
+      ignoreRealInput((event: KeyboardEvent) => {
         blockEnterIfNeeded(event);
 
         if (!isEnabled()) {
@@ -2427,16 +2458,18 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
         if (tryHandleBareKeyShortcut(event)) return;
         if (tryHandleTypeToEdit(event)) return;
 
-        if (!didWindowJustRegainFocus) {
+        // In scoped mode activation is driven by the in-container toolbar, not a
+        // global hotkey, so the host page keeps native Cmd/Ctrl+C behavior.
+        if (!didWindowJustRegainFocus && !getScopeContainer() && !isDemoMode) {
           handleActivationKeys(event);
         }
-      },
+      }),
       { capture: true },
     );
 
     eventListenerManager.addWindowListener(
       "keyup",
-      (event: KeyboardEvent) => {
+      ignoreRealInput((event: KeyboardEvent) => {
         if (blockEnterIfNeeded(event)) return;
 
         if (isSpaceActivationKey(event) && isDragRepositioning()) {
@@ -2550,7 +2583,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
             deactivateRenderer();
           }
         }
-      },
+      }),
       { capture: true },
     );
 
@@ -2566,7 +2599,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
 
     eventListenerManager.addWindowListener(
       "pointermove",
-      (event: PointerEvent) => {
+      ignoreRealInput((event: PointerEvent) => {
         if (!event.isPrimary) return;
         const isTouchPointer = event.pointerType === "touch";
         actions.setTouchMode(isTouchPointer);
@@ -2590,13 +2623,13 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
           clearArrowNavigation();
         }
         handlePointerMove(event.clientX, event.clientY, event.shiftKey);
-      },
+      }),
       { passive: true },
     );
 
     eventListenerManager.addWindowListener(
       "pointerdown",
-      (event: PointerEvent) => {
+      ignoreRealInput((event: PointerEvent) => {
         if (event.button !== 0) return;
         if (!event.isPrimary) return;
         actions.setTouchMode(event.pointerType === "touch");
@@ -2634,13 +2667,13 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
           event.preventDefault();
           event.stopImmediatePropagation();
         }
-      },
+      }),
       { capture: true },
     );
 
     eventListenerManager.addWindowListener(
       "pointerup",
-      (event: PointerEvent) => {
+      ignoreRealInput((event: PointerEvent) => {
         if (event.button !== 0) return;
         if (!event.isPrimary) return;
         if (isEventFromOverlay(event, "data-react-grab-ignore-events")) return;
@@ -2652,13 +2685,13 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
           event.preventDefault();
           event.stopImmediatePropagation();
         }
-      },
+      }),
       { capture: true },
     );
 
     eventListenerManager.addWindowListener(
       "contextmenu",
-      (event: MouseEvent) => {
+      ignoreRealInput((event: MouseEvent) => {
         if (!isRendererActive() || isCopying() || isPromptMode()) return;
         if (editMode.isOpen()) return;
 
@@ -2699,7 +2732,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
         actions.setPointer(position);
         actions.freeze();
         openContextMenu(element, position);
-      },
+      }),
       { capture: true },
     );
 
@@ -2710,7 +2743,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
 
     eventListenerManager.addWindowListener(
       "click",
-      (event: MouseEvent) => {
+      ignoreRealInput((event: MouseEvent) => {
         if (isEventFromOverlay(event, "data-react-grab-ignore-events")) return;
         if (isModalPopoverOpen()) return;
 
@@ -2726,7 +2759,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
             }
           }
         }
-      },
+      }),
       { capture: true },
     );
 
@@ -2919,14 +2952,21 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       grabbedBoxTimeouts.clear();
       labelController.cancelAllFades();
       autoScroller.stop();
-      document.body.style.userSelect = "";
-      document.body.style.touchAction = "";
+      setHostBodyStyle("userSelect", "");
+      setHostBodyStyle("touchAction", "");
       setCursorOverride(null);
       keyboardClaimer.restore();
+      setRuntimeMode(null, false);
     });
 
     const resolvedCssText = typeof cssText === "string" ? cssText : "";
-    const { root: rendererRoot, host: rendererHost } = mountRoot(resolvedCssText);
+    // Demo mode is display-only: nothing inside the overlay should intercept the
+    // host page's real clicks or cursor (the showcase is driven via the API), so
+    // make the entire shadow overlay click-through.
+    const overlayCssText = isDemoMode
+      ? `${resolvedCssText}\n* { pointer-events: none !important; }`
+      : resolvedCssText;
+    const { root: rendererRoot, host: rendererHost } = mountRoot(overlayCssText);
 
     const themeWatcher = watchAppTheme(rendererHost);
     onCleanup(themeWatcher.cleanup);
@@ -3603,6 +3643,19 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       return await copyResolvedElements(elementsArray);
     };
 
+    const resetAll = () => {
+      forceDeactivateAll();
+      dismissAllPopups();
+      actions.clearGrabbedBoxes();
+      actions.clearLabelInstances();
+      actions.setSelectionSource(null, null);
+
+      saveToolbarState(DEFAULT_TOOLBAR_STATE);
+      setCurrentToolbarState(DEFAULT_TOOLBAR_STATE);
+      if (!isEnabled()) setIsEnabled(true);
+      toolbarStateChangeCallbacks.forEach((callback) => callback(DEFAULT_TOOLBAR_STATE));
+    };
+
     const api: ReactGrabAPI = {
       activate: () => {
         actions.setPendingCommentMode(false);
@@ -3662,6 +3715,7 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
           toolbarStateChangeCallbacks.delete(callback);
         };
       },
+      reset: resetAll,
       dispose: () => {
         disposed = true;
         hasInited = false;
