+++ packages/react-grab/src/utils/freeze-animations.ts
@@ -1,6 +1,7 @@
 import { FROZEN_ELEMENT_ATTRIBUTE } from "../constants.js";
 import { createStyleElement } from "./create-style-element.js";
 import { freezeGsap, unfreezeGsap } from "./freeze-gsap.js";
+import { getScopeContainer } from "./runtime-mode.js";
 
 const FROZEN_STYLES = `
 [${FROZEN_ELEMENT_ATTRIBUTE}],
@@ -117,6 +118,8 @@ const finishAnimations = (animations: Iterable<Animation>): void => {
 };
 
 export const freezeAllAnimations = (elements: Element[]): void => {
+  // Scoped mode must not pause animations on the host page.
+  if (getScopeContainer()) return;
   if (elements.length === 0) return;
   if (areElementsSame(elements, lastInputElements)) return;
 
@@ -159,6 +162,7 @@ const unfreezeAllAnimations = (): void => {
 };
 
 export const freezeAnimations = (elements: Element[]): (() => void) => {
+  if (getScopeContainer()) return () => {};
   if (elements.length === 0) {
     unfreezeAllAnimations();
     return () => {};
@@ -169,6 +173,8 @@ export const freezeAnimations = (elements: Element[]): (() => void) => {
 };
 
 export const freezeGlobalAnimations = (): void => {
+  // Scoped mode must never pause the host page's animations.
+  if (getScopeContainer()) return;
   if (globalAnimationStyleElement) return;
 
   globalAnimationStyleElement = createStyleElement(
